+++ styles/Supplier.js
@@ -112,51 +112,9 @@ export const Container = styled.div`
   }
 `;
 
-export const Categories = styled.div`
-  display: flex;
-  gap: 15px 30px;
-  margin-bottom: 40px;
-  overflow-x: scroll;
-  padding: 10px 15px 15px;
-
-  @media (min-width: 992px) {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 15px 30px;
-  }
-`;
-
-export const CategoryCard = styled('div')(
-  (props) => css`
-    background-color: ${props.theme.colors.alto};
-    border-radius: 10px;
-    box-shadow: 0px 0px 15px rgba(0, 0, 0, 0.25);
-    max-width: 120px;
-    padding: 10px;
-
-    p {
-      color: ${props.theme.colors.stratos};
-      font-family: ${props.theme.fonts.duruSans};
-      font-size: 14px;
-      font-style: normal;
-      font-weight: bold;
-      line-height: 15px;
-    }
-  `
-);
-
 export const Messages = styled.div`
   position: absolute;
   z-index: 2;
-
-  /* @media (min-width: 992px) {
-    left: 0;
-    margin-left: auto;
-    margin-right: auto;
-    position: absolute;
-    right: 0;
-    width: 400px;
-  } */
 `;
 
 export const Circle = styled('div')(
@@ -172,3 +130,17 @@ export const Circle = styled('div')(
     width: 208px;
   `
 );
+
+export const CategoryMobileContianer = styled.div`
+  @media (min-width: 992px) {
+    display: none;
+  }
+`;
+
+export const CategoryDesktopContianer = styled.div`
+  display: none;
+
+  @media (min-width: 992px) {
+    display: block;
+  }
+`;
